split the rest of the files up
