+++ parser/src/rules/both/print.rs
@@ -1,14 +1,14 @@
 use super::*;
 use crate::traits::print;
-use std::fmt::Write;
 
 impl print::Print for Rule {
 	fn print(&self) -> print::Result {
 		use crate::rules::Rule::{Course, Requirement};
+		use std::fmt::Write;
 
 		let mut output = String::new();
 
-		let pair = self.either.clone();
+		let pair = self.both.clone();
 
 		if self.has_save_rule() {
 			write!(&mut output, "do both of the following:\n\n")?;
@@ -21,18 +21,18 @@ impl print::Print for Rule {
 			(Requirement(a), Requirement(b)) => {
 				write!(
 					&mut output,
-					"complete either the {} or {} requirement",
+					"complete both the {} and {} requirements",
 					a.print()?,
 					b.print()?
 				)?;
 			}
 			(Course(a), Course(b)) => {
-				write!(&mut output, "take either {} or {}", a.print()?, b.print()?)?;
+				write!(&mut output, "take both {} and {}", a.print()?, b.print()?)?;
 			}
 			(Requirement(a), Course(b)) => {
 				write!(
 					&mut output,
-					"complete the {} requirement or take {}",
+					"complete the {} requirement and take {}",
 					a.print()?,
 					b.print()?
 				)?;
@@ -40,35 +40,35 @@ impl print::Print for Rule {
 			(Course(a), Requirement(b)) => {
 				write!(
 					&mut output,
-					"take {} or complete the {} requirement",
+					"take {} and complete the {} requirement",
 					a.print()?,
 					b.print()?
 				)?;
 			}
 			(Course(a), b) => {
-				write!(&mut output, "either take {} or {}", a.print()?, b.print()?)?;
+				write!(&mut output, "both take {} and {}", a.print()?, b.print()?)?;
 			}
 			(Requirement(a), b) => {
 				write!(
 					&mut output,
-					"either complete the {} requirement or {}",
+					"both complete the {} requirement and {}",
 					a.print()?,
 					b.print()?
 				)?;
 			}
 			(a, Course(b)) => {
-				write!(&mut output, "either {} or take {}", a.print()?, b.print()?)?;
+				write!(&mut output, "both {} and take {}", a.print()?, b.print()?)?;
 			}
 			(a, Requirement(b)) => {
 				write!(
 					&mut output,
-					"either {} or complete the {} requirement",
+					"both {} and complete the {} requirement",
 					a.print()?,
 					b.print()?
 				)?;
 			}
 			(a, b) => {
-				write!(&mut output, "either:\n\n- {}\n\n- or {}", a.print()?, b.print()?)?;
+				write!(&mut output, "both:\n\n- {}\n\n- and {}", a.print()?, b.print()?)?;
 			}
 		};
