+++ parser/src/traits/util.rs
@@ -1,0 +1,3 @@
+pub trait Util {
+	fn has_save_rule(&self) -> bool;
+}
