+++ parser/src/traits/validate.rs
@@ -12,11 +12,3 @@ pub trait Validate {
 		Ok(())
 	}
 }
-
-pub trait PrettyPrint {
-	fn print(&self) -> Result<String, std::fmt::Error>;
-}
-
-pub trait RuleTools {
-	fn has_save_rule(&self) -> bool;
-}
