+++ parser/src/util/mod.rs
@@ -1,73 +1,11 @@
-use serde::de::{self, Deserialize, Deserializer, MapAccess, Visitor};
-use std::error::Error;
-use std::fmt;
-use std::marker::PhantomData;
-use std::str::FromStr;
-use void::Void;
+mod oxford;
+pub use oxford::Oxford;
 
-pub trait Oxford {
-	fn oxford(&self, trailer: &str) -> String;
-}
+mod errors;
+pub use errors::{ParseError, ValidationError};
 
-impl Oxford for Vec<String> {
-	fn oxford(&self, trailer: &str) -> String {
-		if self.len() == 1 {
-			return format!("{}", self[0]);
-		}
-
-		if self.len() == 2 {
-			return format!("{} {} {}", self[0], trailer, self[1]);
-		}
-
-		if let Some((last, rest)) = self.split_last() {
-			return format!("{}, {} {}", rest.join(", "), trailer, last);
-		}
-
-		String::new()
-	}
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ParseError {
-	InvalidAction,
-	UnknownOperator,
-	InvalidValue,
-	UnknownCommand,
-	InvalidVariableName,
-}
-
-impl fmt::Display for ParseError {
-	fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-		fmt.write_str(self.description())
-	}
-}
-
-impl Error for ParseError {
-	fn description(&self) -> &str {
-		"invalid do: command syntax"
-	}
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ValidationError {
-	GivenAreasMustOutputAreas,
-}
-
-impl fmt::Display for ValidationError {
-	fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-		fmt.write_str(self.description())
-	}
-}
-
-impl Error for ValidationError {
-	fn description(&self) -> &str {
-		match &self {
-			ValidationError::GivenAreasMustOutputAreas => {
-				"in a `given: areas of study` block, `what:` must also be `areas of study`"
-			}
-		}
-	}
-}
+mod serde;
+pub use self::serde::{serde_false, string_or_struct, string_or_struct_parseerror};
 
 pub fn pretty_term(term: &str) -> String {
 	format!("{}", term)
@@ -86,129 +24,9 @@ pub fn expand_year(year: u64, mode: &str) -> String {
 	}
 }
 
-pub fn serde_false() -> bool {
-	false
-}
-
-pub fn string_or_struct<'de, T, D>(deserializer: D) -> Result<T, D::Error>
-where
-	T: Deserialize<'de> + FromStr<Err = Void>,
-	D: Deserializer<'de>,
-{
-	// This is a Visitor that forwards string types to T's `FromStr` impl and
-	// forwards map types to T's `Deserialize` impl. The `PhantomData` is to
-	// keep the compiler from complaining about T being an unused generic type
-	// parameter. We need T in order to know the Value type for the Visitor
-	// impl.
-	struct StringOrStruct<T>(PhantomData<fn() -> T>);
-
-	impl<'de, T> Visitor<'de> for StringOrStruct<T>
-	where
-		T: Deserialize<'de> + FromStr<Err = Void>,
-	{
-		type Value = T;
-
-		fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-			formatter.write_str("string or map")
-		}
-
-		fn visit_str<E>(self, value: &str) -> Result<T, E>
-		where
-			E: de::Error,
-		{
-			Ok(FromStr::from_str(value).unwrap())
-		}
-
-		fn visit_map<M>(self, visitor: M) -> Result<T, M::Error>
-		where
-			M: MapAccess<'de>,
-		{
-			// `MapAccessDeserializer` is a wrapper that turns a `MapAccess`
-			// into a `Deserializer`, allowing it to be used as the input to T's
-			// `Deserialize` implementation. T then deserializes itself using
-			// the entries from the map visitor.
-			Deserialize::deserialize(de::value::MapAccessDeserializer::new(visitor))
-		}
-	}
-
-	deserializer.deserialize_any(StringOrStruct(PhantomData))
-}
-
-pub fn string_or_struct_parseerror<'de, T, D>(deserializer: D) -> Result<T, D::Error>
-where
-	T: Deserialize<'de> + FromStr<Err = ParseError>,
-	D: Deserializer<'de>,
-{
-	// This is a Visitor that forwards string types to T's `FromStr` impl and
-	// forwards map types to T's `Deserialize` impl. The `PhantomData` is to
-	// keep the compiler from complaining about T being an unused generic type
-	// parameter. We need T in order to know the Value type for the Visitor
-	// impl.
-	struct StringOrStruct<T>(PhantomData<fn() -> T>);
-
-	impl<'de, T> Visitor<'de> for StringOrStruct<T>
-	where
-		T: Deserialize<'de> + FromStr<Err = ParseError>,
-	{
-		type Value = T;
-
-		fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-			formatter.write_str("string or map")
-		}
-
-		fn visit_str<E>(self, value: &str) -> Result<T, E>
-		where
-			E: de::Error,
-		{
-			Ok(FromStr::from_str(value).unwrap())
-		}
-
-		fn visit_map<M>(self, visitor: M) -> Result<T, M::Error>
-		where
-			M: MapAccess<'de>,
-		{
-			// `MapAccessDeserializer` is a wrapper that turns a `MapAccess`
-			// into a `Deserializer`, allowing it to be used as the input to T's
-			// `Deserialize` implementation. T then deserializes itself using
-			// the entries from the map visitor.
-			Deserialize::deserialize(de::value::MapAccessDeserializer::new(visitor))
-		}
-	}
-
-	deserializer.deserialize_any(StringOrStruct(PhantomData))
-}
-
-#[cfg(test)]
-mod tests {
-	#[test]
-	fn oxford_len_eq0() {
-		use super::Oxford;
-		assert_eq!("", vec![].oxford("and"));
-	}
-
-	#[test]
-	fn oxford_len_eq1() {
-		use super::Oxford;
-		assert_eq!("A", vec!["A".to_string()].oxford("and"));
-	}
-
-	#[test]
-	fn oxford_len_eq2() {
-		use super::Oxford;
-		assert_eq!("A and B", vec!["A".to_string(), "B".to_string()].oxford("and"));
-		assert_eq!("A or B", vec!["A".to_string(), "B".to_string()].oxford("or"));
-	}
-
-	#[test]
-	fn oxford_len_gt3() {
-		use super::Oxford;
-		assert_eq!(
-			"A, B, and C",
-			vec!["A".to_string(), "B".to_string(), "C".to_string()].oxford("and")
-		);
-		assert_eq!(
-			"A, B, or C",
-			vec!["A".to_string(), "B".to_string(), "C".to_string()].oxford("or")
-		);
+pub fn result_to_option<T, E>(r: Result<T, E>) -> Option<T> {
+	match r {
+		Ok(v) => Some(v),
+		Err(_) => None,
 	}
 }
